Page titles and next page include
The title of each page has been added. If the page has a title it will
now display along side the site title in the browser. Start of next
page include. The include can be placed on any page and the if
statement will update the copy and link accordingly.