FORMSPREE intergration with contact form
Using FORMSPREE to send contact emails to Daniel. If the human has
filled out the form and is sucessfully delivered, they are taken to a
new thank you page. This thank you page let's the human know that the
email was sent and prompts them to go read Daniel's journal or connect
on social media.